+++ src/features/related-files/related-files-config.ts
@@ -8,6 +8,8 @@ const defaultUseExcludes = true
 const defaultUseStricterQuickOpenQuery = false
 const defaultUseGlobalSearch = false
 const defaultViewRenderMode = 'compact'
+const defaultMaxLabelLength = 120
+const defaultCollapsedIndicator = '⸱⸱⸱'
 
 export class RelatedFilesConfig extends FeatureConfig {
   private _customExcludes: Record<string, unknown> = {}
@@ -16,6 +18,8 @@ export class RelatedFilesConfig extends FeatureConfig {
   private _useGlobalSearch: boolean = defaultUseGlobalSearch
   private _viewRenderMode: ViewRenderMode = defaultViewRenderMode
   private _hiddenWorkspaceFoldersInGlobalSearch: string[] = []
+  private _maxLabelLength: number = defaultMaxLabelLength
+  private _collapsedIndicator: string = defaultCollapsedIndicator
 
   constructor() {
     super('RelatedFiles')
@@ -29,6 +33,8 @@ export class RelatedFilesConfig extends FeatureConfig {
     const useGlobalSearch = config.get<boolean>('relatedFiles.useGlobalSearch', defaultUseGlobalSearch)
     const viewRenderMode = config.get<ViewRenderMode>('relatedFiles.viewRenderMode', defaultViewRenderMode)
     const hiddenWorkspaceFoldersInGlobalSearch = config.get<string[]>('relatedFiles.hiddenWorkspaceFoldersInGlobalSearch', [])
+    const maxLabelLength = config.get<number>('relatedFiles.maxLabelLength', defaultMaxLabelLength)
+    const collapsedIndicator = config.get<string>('relatedFiles.collapsedIndicator', defaultCollapsedIndicator)
 
     let hasChanged = false
 
@@ -39,6 +45,8 @@ export class RelatedFilesConfig extends FeatureConfig {
       || this._useGlobalSearch !== useGlobalSearch
       || this._viewRenderMode !== viewRenderMode
       || !areArraysShallowEqual(this._hiddenWorkspaceFoldersInGlobalSearch, hiddenWorkspaceFoldersInGlobalSearch)
+      || this._maxLabelLength !== maxLabelLength
+      || this._collapsedIndicator !== collapsedIndicator
     ) {
       this._customExcludes = customExcludes
       this._useExcludes = useExcludes
@@ -46,11 +54,13 @@ export class RelatedFilesConfig extends FeatureConfig {
       this._useGlobalSearch = useGlobalSearch
       this._viewRenderMode = viewRenderMode
       this._hiddenWorkspaceFoldersInGlobalSearch = hiddenWorkspaceFoldersInGlobalSearch
+      this._maxLabelLength = maxLabelLength
+      this._collapsedIndicator = collapsedIndicator
 
       hasChanged = true
     }
 
-    console.debug('[RelatedFiles] Config has been loaded', { hasChanged, customExcludes, useExcludes, useStricterQuickOpenQuery, useGlobalSearch, viewRenderMode })
+    console.debug('[RelatedFiles] Config has been loaded', { hasChanged, customExcludes, useExcludes, useStricterQuickOpenQuery, useGlobalSearch, viewRenderMode, maxLabelLength, collapsedIndicator })
 
     return hasChanged
   }
@@ -87,6 +97,14 @@ export class RelatedFilesConfig extends FeatureConfig {
 
   getUseStricterQuickOpenQuery() {
     return this._useStricterQuickOpenQuery
+  }
+
+  getMaxLabelLength() {
+    return this._maxLabelLength
+  }
+
+  getCollapsedIndicator() {
+    return this._collapsedIndicator
   }
 
   setHiddenWorkspaceFoldersInGlobalSearch(value: string[]) {